refactor(lib): keep url handling out of custom errors
Resolving where the browser goes only ever needed the location string, so
it does not belong to the error module. It takes the string now and lives
in its own file with its tests.
